iris_classification/src/iris_django_lapisco/requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.1.2
isort==5.5.4
joblib==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.2
pandas==1.1.2
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
